chore: moved plugin to separate package
