+++ packages/strapi-plugin-product-actions/.eslintrc.js
