+++ packages/strapi-plugin-product-actions/admin/.eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+    extends: [
+        'eslint-config-primer/client'
+    ],
+    parserOptions: {
+      project: './tsconfig.json',
+      tsconfigRootDir: __dirname,
+    }
+}
