+++ packages/strapi-plugin-product-actions/admin/src/components/Initializer/index.tsx
