+++ packages/strapi-plugin-product-actions/admin/src/components/SeriesProductActions/index.tsx
