+++ packages/strapi-plugin-product-actions/package.json
@@ -1,0 +1,99 @@
+{
+  "name": "@primer/strapi-plugin-product-actions",
+  "version": "0.0.1",
+  "description": "Enables product actions for Strapi",
+  "exports": {
+    "./strapi-admin": {
+      "source": "./admin/src/index.ts",
+      "import": "./dist/admin/index.mjs",
+      "require": "./dist/admin/index.js",
+      "default": "./dist/admin/index.js"
+    },
+    "./strapi-server": {
+      "source": "./server/src/index.ts",
+      "import": "./dist/server/index.mjs",
+      "require": "./dist/server/index.js",
+      "default": "./dist/server/index.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "dist/",
+    "strapi-server.js"
+  ],
+  "scripts": {
+    "build": "npm-run-all clean --parallel build:code build:types",
+    "build:code": "rollup -c",
+    "build:types": "run-p build:types:server build:types:admin",
+    "build:types:server": "tsc -p server/tsconfig.build.json ",
+    "build:types:admin": "tsc -p admin/tsconfig.build.json ",
+    "clean": "rimraf ./dist",
+    "lint": "eslint .",
+    "test:front": "cross-env IS_EE=true jest --config ./jest.config.front.js",
+    "test:ts:front": "tsc -p admin/tsconfig.json",
+    "test:unit": "jest",
+    "test:unit:watch": "jest --watch",
+    "watch": "rollup -c -w"
+  },
+  "keywords": [
+    "strapi",
+    "plugin",
+    "product",
+    "actions"
+  ],
+  "author": "Magda Odrowaz-Zelezik <magda.zelezik@gmail.com>",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/your-username/primer.git",
+    "directory": "packages/strapi-plugin-product-actions"
+  },
+  "bugs": {
+    "url": "https://github.com/your-username/primer/issues"
+  },
+  "homepage": "https://github.com/your-username/primer/tree/main/packages/strapi-plugin-product-actions#readme",
+  "peerDependencies": {
+    "@strapi/strapi": "^5.13.1"
+  },
+  "devDependencies": {
+    "@atlaskit/pragmatic-drag-and-drop": "^1.5.2",
+    "@atlaskit/pragmatic-drag-and-drop-hitbox": "^1.0.3",
+    "@rollup/plugin-commonjs": "^28.0.6",
+    "@rollup/plugin-dynamic-import-vars": "^2.1.5",
+    "@rollup/plugin-json": "^6.1.0",
+    "@rollup/plugin-node-resolve": "^16.0.1",
+    "@rollup/plugin-swc": "^0.4.0",
+    "@strapi/admin": "5.16.0",
+    "@strapi/content-manager": "5.16.0",
+    "@strapi/types": "5.16.0",
+    "@types/node": "^20.0.0",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.0.0",
+    "jest": "^29.0.0",
+    "npm-run-all": "^4.1.5",
+    "rimraf": "^5.0.0",
+    "rollup": "^4.44.0",
+    "typescript": "^5.0.0"
+  },
+  "engines": {
+    "node": ">=18.0.0",
+    "npm": ">=6.0.0"
+  },
+  "strapi": {
+    "kind": "plugin",
+    "name": "primer-product-actions",
+    "displayName": "Product Actions",
+    "description": "Enables product actions"
+  },
+  "dependencies": {
+    "@strapi/design-system": "2.0.0-rc.26",
+    "@strapi/icons": "2.0.0-rc.26",
+    "@strapi/utils": "5.16.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "6.22.3",
+    "styled-components": "6.1.8"
+  }
+}
