+++ packages/strapi-plugin-product-actions/server/.eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+    extends: [
+        'eslint-config-primer/server'
+    ],
+    parserOptions: {
+      project: './tsconfig.json',
+      tsconfigRootDir: __dirname,
+    }
+}
