+++ packages/strapi-plugin-product-actions/server/src/controllers/product-series.ts
@@ -1,5 +1,5 @@
 export const productSeries = {
-  async createProducts(ctx: any) {
+  async createProducts(ctx: Context) {
     try {
       const { id } = ctx.params;
       const { count = 1 } = ctx.request.body;
@@ -15,7 +15,7 @@ export const productSeries = {
     }
   },
 
-  async updateProducts(ctx: any) {
+  async updateProducts(ctx: Context) {
     try {
       const { id } = ctx.params;
       const updateData = ctx.request.body;
