+++ packages/strapi-plugin-product-actions/server/src/routes/product-series.ts
