+++ packages/strapi-plugin-product-actions/server/tsconfig.build.json
@@ -1,0 +1,11 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+      "rootDir": "../",
+      "baseUrl": ".",
+      "outDir": "../dist",
+      "noEmit": false
+    },
+    "include": ["./src"],
+    "exclude": ["node_modules", "**/*.test.ts"]
+  }
