+++ packages/strapi-plugin-product-actions/server/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "@strapi/typescript-utils/tsconfigs/server",
+  "compilerOptions": {
+    "rootDir": "../",
+    "baseUrl": ".",
+  },
+  "include": [
+    "./src"
+  ],
+  "exclude": ["node_modules"]
+}
